+++ packages/react-hooks/src/ConnectionProvider.tsx
@@ -15,71 +15,45 @@ import {
   addRecord,
 } from './recordUtils'
 
-export type useConnectionsOptions = {
-  excludedTypes?: [ConnectionType | string]
+export type UseConnectionsOptions = {
+  excludedTypes?: Array<ConnectionType | string>
   connectionState?: DidExchangeState
 }
 
 const ConnectionContext = createContext<RecordsState<ConnectionRecord> | undefined>(undefined)
 
 /**
- * This method retreives the connection context for the current agent.
+ * This method retrieves the connection context for the current agent.
  * From this you can access all connection records for the agent.
  * @param options options for useConnections hook, lets us filter out specific types and limit states
  * @returns a connection context containing information about the agents connections
  */
-export const useConnections = (options: useConnectionsOptions = {}) => {
+export const useConnections = (options: UseConnectionsOptions = {}) => {
   const connectionContext = useContext(ConnectionContext)
-  if (!connectionContext) {
-    throw new Error('useConnections must be used within a ConnectionContextProvider')
-  }
 
-  let connections = connectionContext.records
+  let connections = connectionContext?.records
+  connections = useMemo(() => {
+    if (!connections) {
+      throw new Error('useConnections must be used within a ConnectionContextProvider')
+    }
+    // do not filter if not filter options are provided to save on a loop
+    if (!options.connectionState && !options.excludedTypes) return connections
 
-  // Only run this code if we don't have options.excludedTypes to limit ammount of loops
-  if (options.connectionState && !options.excludedTypes) {
-    const filteredByState = useMemo(
-      () => connections.filter((c: ConnectionRecord) => c.state === options.connectionState),
-      [connections, options.connectionState]
-    )
-    connections = filteredByState
-  }
+    return connections.filter((record: ConnectionRecord) => {
+      // By default we include this connection
+      // Filter by state (if connectionState is defined)
+      if (options.connectionState && options.connectionState !== record.state) return false
 
-  if (options.excludedTypes) {
-    const filteredByType = useMemo(
-      () =>
-        connections.filter((record: ConnectionRecord) => {
-          const recordTypes = (record.getTag('connectionType') as [string] | null) || []
-          // Filter out by state if needed
-          if (options.connectionState && options.connectionState !== record.state) {
-            return false
-          }
-          for (const type of recordTypes) {
-            if (options.excludedTypes?.includes(type)) return false
-          }
-          return true
-        }),
-      [connections, options.excludedTypes]
-    )
-    connections = filteredByType
-  }
+      // Exclude records with certain connection types (if defined)
+      const recordTypes = record.connectionTypes as Array<ConnectionType | string> | null
+      if (options.excludedTypes && recordTypes) {
+        return recordTypes.some((connectionType) => !options.excludedTypes?.includes(connectionType))
+      }
+      return true
+    })
+  }, [connections, options.connectionState, options.excludedTypes])
 
   return { ...connectionContext, records: connections }
-}
-
-export const useConnectionsByType = (type: [ConnectionType | string]) => {
-  const { records: connections } = useConnections()
-  const filteredConnections = useMemo(
-    () =>
-      connections.filter((record: ConnectionRecord) => {
-        const recordTypes = record.getTag('connectionType') as [string]
-        for (const t of type) {
-          if (recordTypes?.includes(t)) return true
-        }
-      }),
-    [connections, type]
-  )
-  return filteredConnections
 }
 
 export const useConnectionById = (id: string): ConnectionRecord | undefined => {
@@ -87,11 +61,38 @@ export const useConnectionById = (id: string): ConnectionRecord | undefined => {
   return connections.find((c: ConnectionRecord) => c.id === id)
 }
 
-export const useConnectionByState = (state: DidExchangeState): ConnectionRecord[] => {
+/**
+ * @deprecated filtering by state can now be done by passing options to useConnections hook
+ */
+export const useConnectionByState = (state: DidExchangeState | DidExchangeState[]): ConnectionRecord[] => {
+  const states = useMemo(() => (typeof state === 'string' ? [state] : state), [state])
+
   const { records: connections } = useConnections()
+
   const filteredConnections = useMemo(
-    () => connections.filter((c: ConnectionRecord) => c.state === state),
-    [connections, state]
+    () =>
+      connections.filter((r: ConnectionRecord) => {
+        if (states.includes(r.state)) return r
+      }),
+    [connections]
+  )
+  return filteredConnections
+}
+
+/**
+ * @deprecated filtering by state can now be done by passing options to useConnections hook
+ */
+export const useConnectionNotInState = (state: DidExchangeState | DidExchangeState[]): ConnectionRecord[] => {
+  const states = useMemo(() => (typeof state === 'string' ? [state] : state), [state])
+
+  const { records: connections } = useConnections()
+
+  const filteredConnections = useMemo(
+    () =>
+      connections.filter((r: ConnectionRecord) => {
+        if (!states.includes(r.state)) return r
+      }),
+    [connections]
   )
   return filteredConnections
 }