+++ packages/rest/src/controllers/did/DidController.ts
@@ -1,0 +1,37 @@
+import type { DidResolutionResultProps } from '../types'
+
+import { Agent } from '@aries-framework/core'
+import { Controller, Example, Get, Path, Route, Tags } from 'tsoa'
+import { injectable } from 'tsyringe'
+
+import { Did, DidRecordExample } from '../examples'
+
+@Tags('Dids')
+@Route('/dids')
+@injectable()
+export class DidController extends Controller {
+  private agent: Agent
+
+  public constructor(agent: Agent) {
+    super()
+    this.agent = agent
+  }
+
+  /**
+   * Resolves did and returns did resolution result
+   * @param did Decentralized Identifier
+   * @returns DidResolutionResult
+   */
+  @Example<DidResolutionResultProps>(DidRecordExample)
+  @Get('/:did')
+  public async getDidRecordByDid(@Path('did') did: Did) {
+    const resolveResult = await this.agent.dids.resolve(did)
+
+    if (!resolveResult.didDocument) {
+      this.setStatus(500)
+      return { resolveResult }
+    }
+
+    return { ...resolveResult, didDocument: resolveResult.didDocument.toJSON() }
+  }
+}