+++ packages/rest/src/schemas/CredentialOfferRequest.ts
@@ -1,8 +1,0 @@
-import { IsString } from 'class-validator'
-
-import { CredentialOfferTemp } from './CredentialOfferTemplate'
-
-export class CredentialOfferRequest extends CredentialOfferTemp {
-  @IsString()
-  public connectionId!: string
-}