+++ packages/rest/src/schemas/ProofPresentationRequest.ts
@@ -1,8 +1,0 @@
-import { IsString } from 'class-validator'
-
-import { ProofRequestTemplate } from './ProofRequestTemplate'
-
-export class ProofPresentationRequest extends ProofRequestTemplate {
-  @IsString()
-  public connectionId!: string
-}