+++ packages/rest/tests/basicMessage.test.ts
@@ -1,45 +1,48 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import type { Agent, BasicMessageRecord, ConnectionRecord } from '@aries-framework/core'
-import type { Express } from 'express'
+import type { Server } from 'net'
 
+import { BasicMessageEventTypes } from '@aries-framework/core'
 import request from 'supertest'
+import WebSocket from 'ws'
 
-import { setupServer } from '../src/server'
+import { startServer } from '../src'
 
 import { getTestAgent, objectToJson } from './utils/helpers'
 
 describe('BasicMessageController', () => {
-  let app: Express
-  let agent: Agent
-  let connection: ConnectionRecord
+  let server: Server
+  let aliceAgent: Agent
+  let bobAgent: Agent
+  let bobConnectionToAlice: ConnectionRecord
 
   beforeAll(async () => {
-    agent = await getTestAgent('rest agent basic message test', 3011)
-    app = await setupServer(agent, { port: 3000 })
-    const { invitation } = await agent.connections.createConnection()
-    connection = await agent.connections.receiveInvitation(invitation)
+    aliceAgent = await getTestAgent('Basic Message REST Agent Test Alice', 3002)
+    bobAgent = await getTestAgent('Basic Message REST Agent Test Bob', 5034)
+    server = await startServer(bobAgent, { port: 5033 })
+
+    const { outOfBandInvitation } = await aliceAgent.oob.createInvitation()
+    const { outOfBandRecord: bobOOBRecord } = await bobAgent.oob.receiveInvitation(outOfBandInvitation)
+
+    const [bobConnectionAtBobAlice] = await bobAgent.connections.findAllByOutOfBandId(bobOOBRecord.id)
+    bobConnectionToAlice = await bobAgent.connections.returnWhenIsConnected(bobConnectionAtBobAlice!.id)
   })
 
-  describe('get basic messages', () => {
-    test('should return list of basic messages filtered by connectionId', async () => {
-      const spy = jest.spyOn(agent.basicMessages, 'findAllByQuery')
-      const getResult = (): Promise<BasicMessageRecord[]> => spy.mock.results[0].value
-
-      const response = await request(app).get(`/basic-messages/${connection.id}`)
-      const result = await getResult()
-
-      expect(response.statusCode).toBe(200)
-      expect(response.body).toEqual(result.map(objectToJson))
-    })
+  afterEach(() => {
+    jest.clearAllMocks()
   })
 
-  describe('send basic message to connection', () => {
-    test('should give 204 no content when message is send', async () => {
-      const response = await request(app).post(`/basic-messages/${connection.id}`).send({ content: 'Hello!' })
+  describe('Send basic message to connection', () => {
+    test('should give 204 no content when message is sent', async () => {
+      const response = await request(server)
+        .post(`/basic-messages/${bobConnectionToAlice?.id}`)
+        .send({ content: 'Hello!' })
 
       expect(response.statusCode).toBe(204)
     })
+
     test('should give 404 not found when connection is not found', async () => {
-      const response = await request(app)
+      const response = await request(server)
         .post(`/basic-messages/aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa`)
         .send({ content: 'Hello!' })
 
@@ -47,8 +50,46 @@ describe('BasicMessageController', () => {
     })
   })
 
+  describe('Basic Message WebSocket event', () => {
+    test('should return basic message event sent from test agent to clients', async () => {
+      const client = new WebSocket('ws://localhost:5033')
+
+      const waitForMessagePromise = new Promise((resolve) => {
+        client.on('message', (data) => {
+          const event = JSON.parse(data as string)
+
+          if (event.type === BasicMessageEventTypes.BasicMessageStateChanged) {
+            expect(event.payload.basicMessageRecord.connectionId).toBe(bobConnectionToAlice.id)
+            client.terminate()
+            resolve(undefined)
+          }
+        })
+      })
+
+      await request(server).post(`/basic-messages/${bobConnectionToAlice?.id}`).send({ content: 'Hello!' })
+
+      await waitForMessagePromise
+    })
+  })
+
+  describe('Get basic messages', () => {
+    test('should return list of basic messages filtered by connection id', async () => {
+      const spy = jest.spyOn(bobAgent.basicMessages, 'findAllByQuery')
+      const getResult = (): Promise<BasicMessageRecord[]> => spy.mock.results[0].value
+
+      const response = await request(server).get(`/basic-messages/${bobConnectionToAlice.id}`)
+      const result = await getResult()
+
+      expect(response.statusCode).toBe(200)
+      expect(response.body).toEqual(result.map(objectToJson))
+    })
+  })
+
   afterAll(async () => {
-    await agent.shutdown()
-    await agent.wallet.delete()
+    await aliceAgent.shutdown()
+    await aliceAgent.wallet.delete()
+    await bobAgent.shutdown()
+    await bobAgent.wallet.delete()
+    server.close()
   })
 })